Have AI take controller as an argument
This is a structural change that shifts from the AI class creating (via
MoveSet) its own controller to accepting a controller and associating
that with a moveset. Since DefensiveAI relies on creating four "bots",
and those bots inherit access to the moveset via the AI superclass, the
controller pointer is passed through several constructors.

Overall this change feels advantageous as it prevents the creation of
multiple, functionally identical controllers and fits into a better
class structure, but several organizational problems still remain.